Behat: Limit to just PHP 8.2 for now
While the idea of finding and testing against all of the PHP versions in range was good, this slightly fell down once WP 6.3 stopped supported PHP 5.6. Ideally we'd have all plugins supporting a higher minimum of PHP (or at least match the minimum that WP core does), but that requires major releases of the VIP plugins.